Add code for a simple positive quadrant graph
Currently the code has a few magic offsets.  I need to get to the point
where the graph is titled, with labels and value text so I can then see
how best to remove the magic offsets because ideally this should come
from label sizes.

Once that is done, we can focus on an underlying data structure for a
graph.  This will then allow creation of formulae to calculate the
intervals for the axes, and thus will allow us to move to a generalised
quadrant graph.